+++ core-ng-test/src/test/java/core/framework/test/TestModule.java
@@ -82,7 +82,7 @@ public class TestModule extends AbstractTestModule {
         site().session().cookie("SessionId", "localhost");
         site().cdn().host("//cdn");
         site().security().contentSecurityPolicy("default-src 'self' https://cdn; img-src 'self' https://cdn data:; object-src 'none'; frame-src 'none';");
-        site().publishAPI(List.of("0.0.0.0/0"));
+        site().allowAPI(List.of("0.0.0.0/0"));
     }
 
     private void configureHTTP() {