+++ core-ng/src/main/java/core/framework/internal/web/service/WebServiceClient.java
@@ -138,7 +138,7 @@ public class WebServiceClient {
         // handle empty body gracefully, e.g. 503 during deployment
         // handle html error message gracefully, e.g. public cloud LB failed to connect to backend
         if (response.body.length > 0 && response.contentType != null && ContentType.APPLICATION_JSON.mediaType.equals(response.contentType.mediaType)) {
-            ErrorResponse error = errorResponse(response);
+            InternalErrorResponse error = errorResponse(response);
             if (error.id != null && error.errorCode != null) {  // use manual validation rather than annotation to keep the flow straightforward and less try/catch, check if valid error response json
                 LOGGER.debug("failed to call remote service, statusCode={}, id={}, severity={}, errorCode={}, remoteStackTrace={}", statusCode, error.id, error.severity, error.errorCode, error.stackTrace);
                 throw new RemoteServiceException(error.message, parseSeverity(error.severity), error.errorCode, parseHTTPStatus(statusCode));
@@ -147,9 +147,9 @@ public class WebServiceClient {
         throw new RemoteServiceException("failed to call remote service, statusCode=" + statusCode, Severity.ERROR, "REMOTE_SERVICE_ERROR", parseHTTPStatus(statusCode));
     }
 
-    private ErrorResponse errorResponse(HTTPResponse response) {
+    private InternalErrorResponse errorResponse(HTTPResponse response) {
         try {
-            return (ErrorResponse) reader.fromJSON(ErrorResponse.class, response.body);
+            return (InternalErrorResponse) reader.fromJSON(InternalErrorResponse.class, response.body);
         } catch (Throwable e) {
             int statusCode = response.statusCode;
             throw new RemoteServiceException("failed to deserialize remote service error response, statusCode=" + statusCode, Severity.ERROR, "REMOTE_SERVICE_ERROR", parseHTTPStatus(statusCode), e);